+++ policyengine_uk_data/datasets/imputations/__init__.py
@@ -4,3 +4,5 @@ from .wealth import *
 from .income import *
 from .capital_gains import *
 from .services import impute_services
+from .salary_sacrifice import impute_salary_sacrifice
+from .student_loans import impute_student_loan_plan